+++ server/src/use-cases/delete-cases/delete-all-tasks.ts
@@ -2,9 +2,9 @@ import { DeleteRepository } from "@src/repositories/delete-repository";
 import { DeletedTaskResponse } from "./delete-task";
 
 export class DeleteAllTasks {
-  constructor(private deleteRepository: DeleteRepository) {}
+  constructor(private deleteRepository: DeleteRepository) { }
 
-  async execute(): Promise<DeletedTaskResponse> {  
-    return { deleteTrash: await this.deleteRepository.deleteAll() };
+  async execute(userId: string): Promise<DeletedTaskResponse> {
+    return { deleteTrash: await this.deleteRepository.deleteAll(userId) };
   }
 }
